+++ brand-dev-java-core/src/main/kotlin/com/branddev/api/services/blocking/BrandServiceImpl.kt
@@ -29,6 +29,7 @@ import com.branddev.api.models.brand.BrandRetrieveParams
 import com.branddev.api.models.brand.BrandRetrieveResponse
 import com.branddev.api.models.brand.BrandSearchParams
 import com.branddev.api.models.brand.BrandSearchResponse
+import java.util.function.Consumer
 
 class BrandServiceImpl internal constructor(private val clientOptions: ClientOptions) :
     BrandService {
@@ -38,6 +39,9 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
     }
 
     override fun withRawResponse(): BrandService.WithRawResponse = withRawResponse
+
+    override fun withOptions(modifier: Consumer<ClientOptions.Builder>): BrandService =
+        BrandServiceImpl(clientOptions.toBuilder().apply(modifier::accept).build())
 
     override fun retrieve(
         params: BrandRetrieveParams,
@@ -93,6 +97,13 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
 
         private val errorHandler: Handler<JsonValue> = errorHandler(clientOptions.jsonMapper)
 
+        override fun withOptions(
+            modifier: Consumer<ClientOptions.Builder>
+        ): BrandService.WithRawResponse =
+            BrandServiceImpl.WithRawResponseImpl(
+                clientOptions.toBuilder().apply(modifier::accept).build()
+            )
+
         private val retrieveHandler: Handler<BrandRetrieveResponse> =
             jsonHandler<BrandRetrieveResponse>(clientOptions.jsonMapper)
                 .withErrorHandler(errorHandler)
@@ -104,6 +115,7 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("brand", "retrieve")
                     .build()
                     .prepare(clientOptions, params)
@@ -131,6 +143,7 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.POST)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("brand", "ai", "query")
                     .body(json(clientOptions.jsonMapper, params._body()))
                     .build()
@@ -159,6 +172,7 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("brand", "transaction_identifier")
                     .build()
                     .prepare(clientOptions, params)
@@ -186,6 +200,7 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.POST)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("brand", "prefetch")
                     .body(json(clientOptions.jsonMapper, params._body()))
                     .build()
@@ -214,6 +229,7 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("brand", "retrieve-by-ticker")
                     .build()
                     .prepare(clientOptions, params)
@@ -241,6 +257,7 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("brand", "naics")
                     .build()
                     .prepare(clientOptions, params)
@@ -268,6 +285,7 @@ class BrandServiceImpl internal constructor(private val clientOptions: ClientOpt
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("brand", "search")
                     .build()
                     .prepare(clientOptions, params)