+++ brand-dev-java-core/src/test/kotlin/com/branddev/api/models/brand/BrandSearchParamsTest.kt
@@ -12,12 +12,23 @@ internal class BrandSearchParamsTest {
     @Disabled("skipped: tests are disabled for the time being")
     @Test
     fun create() {
-        BrandSearchParams.builder().query("query").build()
+        BrandSearchParams.builder().query("query").timeoutMs(1L).build()
     }
 
     @Disabled("skipped: tests are disabled for the time being")
     @Test
     fun queryParams() {
+        val params = BrandSearchParams.builder().query("query").timeoutMs(1L).build()
+
+        val queryParams = params._queryParams()
+
+        assertThat(queryParams)
+            .isEqualTo(QueryParams.builder().put("query", "query").put("timeoutMS", "1").build())
+    }
+
+    @Disabled("skipped: tests are disabled for the time being")
+    @Test
+    fun queryParamsWithoutOptionalFields() {
         val params = BrandSearchParams.builder().query("query").build()
 
         val queryParams = params._queryParams()