+++ brand-dev-java-core/src/test/kotlin/com/branddev/api/services/ServiceParamsTest.kt
@@ -45,6 +45,7 @@ internal class ServiceParamsTest {
                 .domain("domain")
                 .forceLanguage(BrandRetrieveParams.ForceLanguage.ALBANIAN)
                 .maxSpeed(true)
+                .timeoutMs(1L)
                 .putAdditionalHeader("Secret-Header", "42")
                 .putAdditionalQueryParam("secret_query_param", "42")
                 .build()